Changes to deal with BIOM format with QIIME 1.5.0
Made sure that if QIIME 1.5.0 is installed, we are generating a tab delimited format OTU table.
The BIOM format OTU table is used in all QIIME processes, but the tab delimited OTU table is used for all AutoQIIME scripts.
If QIIME 1.4.0 is installed, we simply do what we were doing before.