Added ability to exclude OTUs from OTU table
With <exclude-taxonomy taxa="keyword"/>, you can exclude OTUs with RDP classifications that hit against a comma separated list of keywords.
One valid use of this is if RDP keeps classifying OTUs as "Unclassified" (maybe because they are 18s), you can specify <exclude-taxonomy taxa="Unclassified"/> and those OTUs will not make it into the OTU table, and thus none of the downstream analyses.
Not sure whether to put this in the official documentation or not. It's not the most honest thing to do to your data...