Added a warning for raw-cdhit about bad flags
cd-hit doesn't provide a good way to get the version, so if the user is using cd-hit < 4.5.4, the -M 0 and -T flags cause it to print usage.
Print a little message before the cd-hit-est call warning the user about this, and make sure the cd-hit-est command is printed.
This doesn't affect using cdhit through QIIME.